+++ training/results/summary.py
@@ -226,11 +226,14 @@ benchmark_datasets = {
         "general_info": [
             "Energies and forces for molecular dynamics trajectories of eight organic molecules. ",
             "All geometries in A, energy labels in kcal/mol and force labels in kcal/mol/A. ",
-            "We use preset train-test split. Errors are MAE for forces.",
+            "We use preset train-test split. Errors are MAE for forces. ",
+            "Results are for the CCSD and CCSD(T) data in MD17."
         ],
         "targets": [
-            {"metric": "val_scaled_mean_absolute_error", "name": "Aspirin", "find_best": "min",
-             "trajectory_name": ""},
+            {"metric": "val_force_scaled_mean_absolute_error", "name": "Aspirin", "find_best": "min",
+             "trajectory_name": "aspirin_ccsd"},
+            {"metric": "val_force_scaled_mean_absolute_error", "name": "Toluene", "find_best": "min",
+             "trajectory_name": "toluene_ccsd_t"}
         ]
     },
 }