+++ changelog.md
@@ -3,6 +3,8 @@ v2.2.0
 * Removed deprecated modules in ``kgcnn.utils``.
 * Moved ``kgcnn.utils.model`` to `kgcnn.model.utils`.
 * Fixed behaviour for ``kgcnn.data.base.MemoryGraphDataset.get_train_test_indices`` to return list of train-test index tuples.
+* Updated ``kgcnn.hyper.hyper.HyperParameter`` to deserialize metrics and loss for multi-output models.
+* Added `trajectory_name` in `summary.py` and `history.py`. 
 
 
 v2.1.1